Added ability to define external img tools.
This allows img to be extended with other shellscripts/go/ruby/whatever
as long as it is on the path and called img-something it can be
executed with 'img something args'.

Will define a proper interface next; so for example, help strings can
be added to the 'img help' interface.