Simplified README. Removed lots of contents.
Reason: I think it was trying to do too much. You can easily find
out about the tools by running help, which is going to be the most
up to date.

The README now just gives a broad view of how to use img on the
command line and as a library. The section on External Scripts
may be changed further, in the future.