Parallelised pixelate, and added option to crop.
It isn't the best implementation, but it gets the job done. I've
pulled the code that works on averaging colours out of the main
loop. It now just runs over a list of Rectagles.

Also added --crop option, which does what pixelate used to do and
only show exact pixels. The default is now to show the full image,
that is the output image is the same size as the input.