Tentatively adding partial exif support
This is probably very buggy, and requires you to explicitly set the
output to --jpeg, but does seem to be working in the limited tests
I've run.

The dream, of course, would be to get TIFF files to support exif,
but it seems to be causing a few issues. That is, the image won't
display. Major.

For now this will do.