Fix chaining multiple commands together
Instead of it failing outright the error (which is trying to load the EXIF data
in the next image in the chain) it now silently fails.

This means that attempting to chain multiple commands and copy the EXIF by
setting all to JPEG will write the image correctly with no EXIF data.